requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.3.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
braceexpand==0.1.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dateparser==1.0.0
decorator==5.0.7
discord==1.0.1
discord-ext-menus==1.0
discord.py==1.7.1
Django==3.2
ExtractTable==2.1.2
geographiclib==1.50
geopy==2.1.0
humanize==3.4.1
idna==3.1
import-expression==1.1.4
jishaku==1.20.0.220
llvmlite==0.36.0
multidict==5.1.0
mypy-extensions==0.4.3
numba==0.53.1
numpy==1.20.2
pandas==1.2.3
pathspec==0.8.1
pendulum==2.1.2
Pillow==8.2.0
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
pytzdata==2020.1
regex==2021.4.4
requests==2.25.1
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
timezonefinder==5.2.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
utils==1.0.1
validators==0.18.2
watchgod==0.7
